Fix xorshift_rand seeding

+++ eval7/xorshift_rand.pyx
@@ -71,7 +71,7 @@ def init(seed=None):
     del time
     del struct
 
-    cy_init(seed)
+    cy_init(long(seed))
 
 
 def randint(int n):
